backend_django/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
dj-rest-auth>=5.0,<6.0
django-allauth>=0.61,<1.0

# PDF Generation
PyMuPDF>=1.24,<2.0
reportlab>=4.1,<5.0

# Data Processing
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0

# Cloud Storage
google-cloud-storage>=2.14,<3.0

# Google Auth (for OAuth token verification)
google-auth>=2.29,<3.0

# AI (Fase 6 - optional)
# anthropic>=0.18,<1.0
# openai>=1.12,<2.0
# pgvector>=0.2,<1.0

# Utilities
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Development
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
flake8>=7.0,<8.0
black>=24.1,<25.0

# Background tasks (optional - for async operations)
# celery>=5.3,<6.0
# redis>=5.0,<6.0
